# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GLPI Dashboard API
  description: |
    API para o Dashboard GLPI - Sistema de monitoramento e métricas de tickets.
    
    Esta API fornece endpoints para:
    - Métricas de tickets (novos, pendentes, em progresso, resolvidos)
    - Ranking de técnicos
    - Alertas do sistema
    - Estatísticas de performance
    - Health checks
    
    ## Autenticação
    A API utiliza tokens de autenticação do GLPI. Configure as variáveis de ambiente:
    - `GLPI_URL`: URL do servidor GLPI
    - `GLPI_USER_TOKEN`: Token do usuário
    - `GLPI_APP_TOKEN`: Token da aplicação
    
    ## Rate Limiting
    - 100 requests por minuto por IP
    - 1000 requests por hora por token
    
    ## Cache
    - Métricas: cache de 5 minutos
    - Rankings: cache de 15 minutos
    - Status: sem cache
  version: 1.0.0
  contact:
    name: Equipe de Desenvolvimento
    email: dev@empresa.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desenvolvimento
  - url: https://glpi-dashboard.empresa.com/api
    description: Servidor de produção
paths:
...
  /status:
    get:
      summary: Obter status do sistema
      description: Retorna status detalhado do sistema e serviços
      operationId: getSystemStatus
      tags:
        - Sistema
      responses:
        '200':
          description: Status do sistema
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      api:
                        type: string
                        enum: [online, offline]
                      glpi:
                        type: string
                        enum: [online, offline, error]
                      glpi_message:
                        type: string
                      glpi_response_time:
                        type: number
                      last_update:
                        type: string
                        format: date-time
                      version:
                        type: string
                      uptime:
                        type: string
                    required:
                      - api
                      - glpi
                      - glpi_message
                      - glpi_response_time
                      - last_update
                      - version
                      - uptime
                  overall_status:
                    type: string
                    enum: [healthy, degraded, unhealthy]
                  response_time_ms:
                    type: number
                required:
                  - success
                  - data
                  - overall_status
                  - response_time_ms
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /filter-types:
    get:
      summary: Obter tipos de filtro disponíveis
      description: Retorna lista de tipos de filtro suportados pela API
      operationId: getFilterTypes
      tags:
        - Configuração
      responses:
        '200':
          description: Tipos de filtro disponíveis
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      creation:
                        type: object
                        properties:
                          name:
                            type: string
                          description:
                            type: string
                          default:
                            type: boolean
                        required:
                          - name
                          - description
                          - default
                      modification:
                        type: object
                        properties:
                          name:
                            type: string
                          description:
                            type: string
                          default:
                            type: boolean
                        required:
                          - name
                          - description
                          - default
                      current_status:
                        type: object
                        properties:
                          name:
                            type: string
                          description:
                            type: string
                          default:
                            type: boolean
                        required:
                          - name
                          - description
                          - default
                    required:
                      - creation
                      - modification
                      - current_status
                  message:
                    type: string
                required:
                  - success
                  - data
                  - message
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
...
  schemas:
...
    ApiError:
      type: object
      description: Schema para erros da API
      properties:
        success:
          type: boolean
          default: false
          description: Sempre false para erros
        message:
          type: string
          description: Mensagem de erro
        errors:
          type: array
          items:
            type: string
          description: Lista detalhada de erros
        timestamp:
          type: string
          format: date-time
          description: Timestamp do erro
        error_code:
          type: string
          nullable: true
          description: Código de erro específico
      required:
        - success
        - message
        - errors
        - timestamp
...